Attempt to fix sizing issue with 1st image unsuccessfully
Tried to fix the issue with the first image's size being too small
and lining up with the other 4 images. The first image should be
bigger than the others. Tried inline CSS and adding another div
class to the other 4 elements in the HTML and changing the CSS
to reflect that. Neither worked.